backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
httpx
pydantic
